+++ KnobScripter/info.py
@@ -1,0 +1,6 @@
+""" Version/author information """
+__version__ = "3.0a0"
+__author__ = "Adrian Pueyo"
+__date__ = "Dec 11 2020"
+__mail__ = "pueyovfx@gmail.com"
+__web__ = "www.adrianpueyo.com"
